+++ CHANGELOG.md
@@ -5,8 +5,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Preferences window
+- option to autostart LaterOn on loggin in (MacOS, Windows)
+
 ### Changed
 - improve UI
+- show LaterOn on MacOS's dock
 
 ## [0.0.4] - 2021-04-22
 ### Changed