+++ CHANGELOG.md
@@ -5,6 +5,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+## [0.1.0] - 2021-04-22
 ### Added
 - Preferences window
 - option to autostart LaterOn on loggin in (MacOS, Windows)
@@ -43,7 +44,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - tray menu to open config file
 - `cron`, `once` and `repeat` reminders
 
-[Unreleased]: https://github.com/hovancik/later-on/compare/v0.0.4...HEAD
+[Unreleased]: https://github.com/hovancik/later-on/compare/v0.1.0...HEAD
+[0.1.0]: https://github.com/hovancik/later-on/releases/tag/v0.1.0
 [0.0.4]: https://github.com/hovancik/later-on/releases/tag/v0.0.4
 [0.0.3]: https://github.com/hovancik/later-on/releases/tag/v0.0.3
 [0.0.2]: https://github.com/hovancik/later-on/releases/tag/v0.0.2